Sprinkle Works p1
Made the sprinkler capable of moving around
Made the sprinkler follow the camera by childing, either the sprinkler is childed to camera or not.
Tried to make the sprinkler detect Tiles and shorten the time while returning when it gets out. (It can shorten the time but not lengthen when going out of range)